events_generator/requirements.txt
kafka-python==2.1.0
numpy==1.24.4
python-dotenv==1.0.0
